+++ src/github_user/entrypoints/api.py
@@ -17,6 +17,8 @@ from src.github_user.entrypoints.schema import (
     GithubUserRenewAllResponseDto,
     GithubUserRenewOneRequestDto,
     GithubUserRenewOneResponseDto,
+    SNSGithubCallbackRequestDto,
+    SNSGithubCallbackResponseDto,
 )
 from src.github_user.adapters.external_api import RequestExternalAPIClient
 from src.github_user.adapters.repository import GithubUserRepository
@@ -25,10 +27,13 @@ from src.github_user.services.use_cases.list_github_user import GithubUserListUs
 from src.github_user.services.use_cases.approve_github_user import GithubUserApproveUseCase
 from src.github_user.services.use_cases.renew_one_github_user import GithubUserRenewOneUseCase
 from src.github_user.services.use_cases.renew_all_github_user import GithubUserRenewAllUseCase
+from src.github_user.services.use_cases.sns_github_login import SNSGithubLoginUseCase
 from src.utils.permissions import IsAdmin, check_permissions
+from src.config import CONFIG
 
 router = APIRouter()
 security = HTTPBearer()
+# TODO : api 구조화
 
 
 @router.post(
@@ -135,6 +140,30 @@ def renew_all_github_user(
     input_dto = GithubUserRenewAllRequestDto()
 
     use_case = GithubUserRenewAllUseCase(repo=repo, crawler=crawler)
+
+    output_dto = use_case.execute(input_dto)
+    return output_dto
+
+
+@router.get("/sns/github")
+def get_github_login_url():
+    url = "https://github.com/login/oauth/authorize?client_id={0}&redirect_uri={1}".format(
+        CONFIG.GITHUB_API_CLIENT_ID, CONFIG.GITHUB_OAUTH_REDIRECT_URI
+    )
+
+    return {"login_url": url}
+
+
+@router.get("/sns/github/callback")
+def github_callback(
+    code: str,
+    session: Session = Depends(get_session),
+) -> SNSGithubCallbackResponseDto:
+    repo = GithubUserRepository(session)
+    external_api = RequestExternalAPIClient()
+    input_dto = SNSGithubCallbackRequestDto(code=code)
+
+    use_case = SNSGithubLoginUseCase(repo=repo, external_api=external_api)
 
     output_dto = use_case.execute(input_dto)
     return output_dto
